requirements.txt
alembic==1.7.7
anyio==3.5.0
apache-airflow==2.2.4
apache-airflow-providers-apache-spark==2.1.1
apache-airflow-providers-ftp==2.1.0
apache-airflow-providers-http==2.1.0
apache-airflow-providers-imap==2.2.1
apache-airflow-providers-sqlite==2.1.1
apache-airflow-upgrade-check==1.4.0
apispec==3.3.2
argcomplete==1.12.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.0
asttokens==2.0.5
async-generator==1.10
attrs==20.3.0
Automat==0.8.0
Babel==2.9.1
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
bleach==4.1.0
blinker==1.4
cached-property==1.5.2
cachelib==0.6.0
cachetools==4.2.4
cattrs==1.10.0
certifi==2019.11.28
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
Click==7.0
clickclick==20.10.2
cloud-init==21.4
colorama==0.4.3
colorlog==4.0.2
command-not-found==0.3
commonmark==0.9.1
configobj==5.0.6
configparser==3.5.3
connexion==2.12.0
constantly==15.1.0
croniter==0.3.37
cryptography==2.9.2
db-dtypes==0.3.1
dbus-python==1.2.16
debugpy==1.5.1
decorator==4.4.2
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.4
distro==1.4.0
distro-info===0.23ubuntu1
Django==4.0.3
django-elasticsearch-dsl==7.2.2
django-elasticsearch-dsl-drf==0.22.4
django-nine==0.2.5
djangorestframework==3.13.1
dnspython==2.2.1
docutils==0.16
elastic-transport==8.1.0
elasticsearch==7.17.1
elasticsearch-dsl==7.4.0
email-validator==1.1.3
entrypoints==0.3
executing==0.8.3
Flask==1.1.4
Flask-Admin==1.5.4
Flask-AppBuilder==3.4.4
Flask-Babel==1.0.0
Flask-Caching==1.10.1
Flask-JWT-Extended==3.25.1
Flask-Login==0.4.1
Flask-OpenID==1.3.0
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
flask-swagger==0.2.13
Flask-WTF==0.14.3
funcsigs==1.0.2
future==0.18.2
google-api-core==2.7.1
google-api-python-client==1.12.10
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.0
google-cloud-bigquery==2.34.2
google-cloud-bigquery-storage==2.13.0
google-cloud-bigtable==2.7.0
google-cloud-container==2.10.6
google-cloud-core==2.2.3
google-cloud-dlp==3.6.2
google-cloud-language==2.4.1
google-cloud-secret-manager==2.9.2
google-cloud-spanner==3.13.0
google-cloud-speech==2.13.1
google-cloud-storage==2.2.0
google-cloud-texttospeech==2.11.0
google-cloud-translate==3.7.2
google-cloud-videointelligence==2.6.1
google-cloud-vision==2.7.1
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.55.0
graphviz==0.19.1
grpc-google-iam-v1==0.12.3
grpcio==1.44.0
grpcio-gcp==0.2.2
grpcio-status==1.44.0
gunicorn==20.1.0
h11==0.12.0
httpcore==0.14.7
httplib2==0.20.4
httpx==0.22.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==4.11.3
importlib-resources==5.4.0
incremental==16.10.1
inflection==0.5.1
ipykernel==6.9.2
ipython==8.1.1
ipython-genutils==0.2.0
ipywidgets==7.7.0
iso8601==1.0.2
itsdangerous==1.1.0
jedi==0.18.1
Jinja2==2.10.1
json-merge-patch==0.2
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.3
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.1.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.7.1
lockfile==0.12.2
Mako==1.2.0
Markdown==2.6.11
MarkupSafe==2.0.1
marshmallow==3.15.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
matplotlib-inline==0.1.3
mistune==0.8.4
more-itertools==4.2.0
mysql-connector-python==8.0.28
natsort==8.1.0
nbclient==0.5.13
nbconvert==6.4.4
nbformat==5.2.0
nest-asyncio==1.5.4
netifaces==0.10.4
nftables==0.1
notebook==6.4.10
numpy==1.22.3
oauthlib==3.1.0
outcome==1.1.0
packaging==21.3
pandas==0.25.3
pandas-gbq==0.17.4
pandocfilters==1.5.0
parso==0.8.3
pendulum==2.1.2
pexpect==4.6.0
pickleshare==0.7.5
prison==0.2.1
prometheus-client==0.13.1
prompt-toolkit==3.0.28
proto-plus==1.20.3
protobuf==3.19.4
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.3
pyarrow==7.0.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.21
pydata-google-auth==1.4.0
Pygments==2.11.2
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyMySQL==1.0.2
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==3.0.7
pyrsistent==0.15.5
pyserial==3.4
PySocks==1.7.1
pyspark==3.2.1
pythainlp==3.0.5
python-apt==2.0.0+ubuntu0.20.4.7
python-daemon==2.3.0
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-nvd3==0.15.0
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2021.3
pytzdata==2020.1
PyYAML==5.3.1
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.1
requests==2.27.1
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
rfc3986==1.5.0
rich==12.0.0
rsa==4.8
SecretStorage==2.3.1
selenium==4.1.3
selinux==3.0
Send2Trash==1.8.0
sentry-sdk==1.5.7
service-identity==18.1.0
setproctitle==1.2.2
simplejson==3.16.0
six==1.14.0
slip==0.6.5
slip.dbus==0.6.5
sniffio==1.2.0
sortedcontainers==2.4.0
sos==4.3
soupsieve==2.3.1
SQLAlchemy==1.3.24
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.38.2
sqlparse==0.4.2
ssh-import-id==5.10
stack-data==0.2.0
statsd==3.3.0
swagger-ui-bundle==0.0.9
systemd-python==234
tabulate==0.8.9
tenacity==8.0.1
termcolor==1.1.0
terminado==0.13.3
testpath==0.6.0
text-unidecode==1.3
thrift==0.15.0
tinydb==4.7.0
tornado==6.1
traitlets==5.1.1
trio==0.20.0
trio-websocket==0.9.2
Twisted==18.9.0
tzlocal==1.5.1
ubuntu-advantage-tools==27.6
ufw==0.36
unattended-upgrades==0.1
unicodecsv==0.14.1
uritemplate==3.0.1
urllib3==1.26.9
wadllib==1.3.3
wcwidth==0.2.5
webdriver-manager==3.5.4
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.6.0
wrapt==1.14.0
wsproto==1.1.0
WTForms==2.3.3
zipp==3.7.0
zope.deprecation==4.4.0
zope.interface==4.7.1
